Working on creating class to handle geometry manuipulation
