--- src/UI/Geometry/2D/Abstraction/geometry2DAbstraction.cpp
@@ -1,29 +1,0 @@
-#include <UI/geometryEditor2D.h>
-
-
-
-int geometryEditor2DAbstraction::getDrawingSizeHeight()
-{
-	return drawingSizeHeight;
-}
-
-
-
-void geometryEditor2DAbstraction::setDrawingSizeHeight(int height)
-{
-	drawingSizeHeight = height;
-}
-
-
-
-int geometryEditor2DAbstraction::getDrawingSizeWidth()
-{
-	return drawingSizeWidth;
-}
-
-
-
-void geometryEditor2DAbstraction::setDrawingSizeWidth(int width)
-{
-	drawingSizeWidth = width;
-}
