--- src/UI/Geometry/2D/Controller/geometry2DController.cpp
@@ -1,29 +1,0 @@
-#include <UI/geometryEditor2D.h>
-
-
-
-int geometryEditor2DController::getDrawingHeightAbstract()
-{
-	return abstraction.getDrawingSizeHeight();
-}
-
-
-
-void geometryEditor2DController::setDrawingHeightAbstract(int height)
-{
-	abstraction.setDrawingSizeHeight(height);
-}
-
-
-
-void geometryEditor2DController::setDrawingWidthAbstract(int width)
-{
-	abstraction.setDrawingSizeWidth(width);
-}
-
-
-
-int geometryEditor2DController::getDrawingWidthAbstract()
-{
-	return abstraction.getDrawingSizeWidth();
-}
