+++ src/UI/Geometry/GeometryEditor2D.cpp
@@ -1,0 +1,36 @@
+#include <UI/GeometryEditor2D.h>
+
+void geometryEditor2D::addNode(double &xPoint, double &yPoint, double &distance)
+{
+    edgeLineShape *edgeLine;
+	edgeLine = new edgeLineShape();
+	arcShape *arcSegment;
+	arcSegment = new arcShape();
+	
+	Vector firstNode, secondNode, thirdNode, c;
+	int nodeListSize = _nodeList.size();
+	node newNode;
+    
+    
+}
+
+
+
+void geometryEditor2D::addBlockLabel()
+{
+    
+}
+
+
+
+void geometryEditor2D::addLine()
+{
+    
+}
+
+
+
+void geometryEditor2D::addArc()
+{
+    
+}
